IOChannelFactory: update expand to include size
This is just a demo. Have not added new coverage other than things like
TextIO tests that actually read files.

A quick survey of three file systems (GCS, S3, and Local) showed that the
"expand glob" APIs used to list files include size as well. We should expose this to
the sources directly! This enables:

* Accurate size estimation, rather than sampling, even for large numbers of files
  in FileBasedSource.
* Elides an extra file stat in FileBasedSource. (We build single-file sources with the
  correct end offset, rather than Long.MAX_VALUE.)

Cleaned up a few other users and the accompanying tests.

Deleted some tests of the now-unused code.